refactor: moved shared filder into kuwala-pipeline + restructured the project context in preps for dockerization, which will help in terms of docker context and all that.
