+++ kuwala-pipelines/osm-poi/prettier.config.js
@@ -1,3 +1,3 @@
-const config = require('../../shared/js/codeStyle/prettier.config');
+const config = require('../shared/js/codeStyle/prettier.config');
 
 module.exports = config;
