+++ kuwala-pipelines/osm-poi/src/app/routes/geojson/index.js
@@ -1,5 +1,5 @@
 const Router = require('express');
-const { Errors, H3Utils } = require('../../../../../../shared/js');
+const { Errors, H3Utils } = require('../../../../../shared/js');
 const { h3Config } = require('../../../../config');
 const { Poi } = require('../../../models');
