+++ kuwala-pipelines/osm-poi/src/app/routes/radius/index.js
@@ -1,6 +1,6 @@
 const Router = require('express');
 const { h3Config } = require('../../../../config');
-const { H3Utils } = require('../../../../../../shared/js');
+const { H3Utils } = require('../../../../../shared/js');
 const { Poi } = require('../../../models');
 
 const router = Router();
