+++ kuwala-pipelines/shared/js/codeStyle/.eslintrc.json
