+++ kuwala-pipelines/shared/js/src/errors/generalError/index.js
